adjustments to auto complete api
