+++ server/api/search/utils/prefixUtils.js
